pyproject.toml
[project]
name = "deribit-api"
version = "0.1.0"
description = "API для получения цен криптовалют с биржи Deribit"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "celery-types==0.24.0",
    "dishka>=1.7.2",
    "fastapi[standard]>=0.128.0",
    "orjson>=3.10.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.45",
]
